lex parser debugging started with flexmark

+++ flexmark-ext-abbreviation/src/main/java/com/vladsch/flexmark/ext/abbreviation/internal/AbbreviationNodePostProcessor.java
@@ -56,17 +56,16 @@ public class AbbreviationNodePostProcessor extends NodePostProcessor {
         BasedSequence literal = Escaping.unescape(original, textMapper);
 
         Matcher m = abbreviations.matcher(literal);
-        Node lastNode = node;
         int lastEscaped = 0;
         boolean wrapInTextBase = !(node.getParent() instanceof TextBase);
-        TextBase textBase = null;
+        TextBase textBase = wrapInTextBase ? null : (TextBase) node.getParent();
 
         while (m.find()) {
             //String found = m.group();
             if (abbreviationMap.containsKey(m.group(0))) {
                 String abbreviation = abbreviationMap.get(m.group(0));
 
-                BasedSequence abbrText = literal.subSequence(m.start(0), m.end(0));
+                BasedSequence toDecorateText = literal.subSequence(m.start(0), m.end(0));
                 int startOffset = textMapper.originalOffset(m.start(0));
                 int endOffset = textMapper.originalOffset(m.end(0));
 
@@ -74,40 +73,39 @@ public class AbbreviationNodePostProcessor extends NodePostProcessor {
                     wrapInTextBase = false;
                     textBase = new TextBase(original);
                     node.insertBefore(textBase);
-                    textBase.appendChild(node);
                     state.nodeAdded(textBase);
                 }
 
                 if (startOffset != lastEscaped) {
                     BasedSequence escapedChars = original.subSequence(lastEscaped, startOffset);
                     Node node1 = new Text(escapedChars);
-                    lastNode.insertAfter(node1);
-                    lastNode = node1;
-                    state.nodeAdded(lastNode);
+                    textBase.appendChild(node1);
+                    state.nodeAdded(node1);
                 }
 
-                BasedSequence origAbbrText = original.subSequence(startOffset, endOffset);
-                Abbreviation abbrNode = new Abbreviation(origAbbrText, abbreviation);
-                lastNode.insertAfter(abbrNode);
-                lastNode = abbrNode;
-                Text node1 = new Text(origAbbrText);
-                abbrNode.appendChild(node1);
-                state.nodeAddedWithChildren(lastNode);
+                BasedSequence origToDecorateText = original.subSequence(startOffset, endOffset);
+                Abbreviation decorationNode = new Abbreviation(origToDecorateText, abbreviation);
+                textBase.appendChild(decorationNode);
+                //Text undecoratedTextNode = new Text(origToDecorateText);
+                //decorationNode.appendChild(undecoratedTextNode);
+                //state.nodeAddedWithChildren(decorationNode);
+                state.nodeAdded(decorationNode);
 
                 lastEscaped = endOffset;
             }
         }
 
-        if (lastEscaped != original.length()) {
-            BasedSequence escapedChars = original.subSequence(lastEscaped, original.length());
-            Node node1 = new Text(escapedChars);
-            lastNode.insertAfter(node1);
-            state.nodeAdded(lastNode);
-        }
+        if (lastEscaped > 0) {
+            if (lastEscaped != original.length()) {
+                BasedSequence escapedChars = original.subSequence(lastEscaped, original.length());
+                Node node1 = new Text(escapedChars);
+                textBase.appendChild(node1);
+                state.nodeAdded(node1);
+            }
 
-        lastNode = node.getNext();
-        node.unlink();
-        state.nodeRemoved(node);
+            node.unlink();
+            state.nodeRemoved(node);
+        }
     }
 
     public static class Factory extends NodePostProcessorFactory {
